add `operator` argument to `find` method; move default sort
A previous commit in this PR added a `case_sensitive` argument to
the `find` helper method. This replaces that argument with `operator`,
defaulting to '=' for backward compatibility. This is more flexible,
and introduces support for not only ILIKE, but also LIKE

The default sort is moved out of the `find` and `keyword_search` methods,
becoming a constant in the `Search` class.