Some updates for the reindex job (#55)
If a doctype doesn't exist the batch job can't be created.
For now just cut back on the doctypes (Osteology is not in
all profiles and removed record types [very likely] not
impacted by indexing).

Reset the reindex job name. If the job has already been
created it cannot be created again with a different name
for whatever reason (which is what the client attempts
to do if the name does not match).

Slightly expanded test coverage for CSpace Response.